requirements-render.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
boto3==1.34.1
botocore==1.34.162
gunicorn==21.2.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
Werkzeug==2.3.7
email-validator==2.0.0
